pyproject.toml
[tool.poetry]
name = "electronics-retail-chain"
version = "0.1.0"
description = ""
authors = ["Sergey Arefev"]
readme = "README.md"
packages = [{include = "electronics_retail_chain"}]

[tool.poetry.dependencies]
python = "^3.11"
django = "4.2"
djangorestframework = "^3.15.2"
psycopg2-binary = "^2.9.9"
python-dotenv = "^1.0.1"
getenv = "^0.2.0"
drf-yasg = "^1.21.7"
djangorestframework-simplejwt = "^5.3.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
